+++ src/device-registry/models/SiteActivity.js
