+++ src/device-registry/utils/does-device-exist.js
