+++ src/device-registry/utils/generate-filter.js
