+++ src/device-registry/utils/insert-measurements.js
