+++ src/device-registry/utils/transform-measurements.js
